requirements.txt
pandas
python-dotenv
cryptography
google-generativeai
